Reorganize sidebar app files
