+++ types/applications/sidebar.d.ts
