+++ types/applications/sidebarTabs/combatTracker.d.ts
