+++ types/applications/sidebarTabs/compendiumDirectory.d.ts
