+++ types/applications/sidebarTabs/sidebarDirectories/clientSettings.d.ts
