+++ types/applications/sidebarTabs/sidebarDirectories/itemDirectory.d.ts
