+++ types/applications/sidebarTabs/sidebarDirectories/journalDirectory.d.ts
