+++ types/applications/sidebarTabs/sidebarDirectory.d.ts
